Switch to the ARC merged binutils-gdb repository.
This commit updates the version of binutils used to be from the merged
binutils-gdb repository.  The default version is a commit from the
arc-2.23-dev branch.

Building of gdb from the merged binutils-gdb repository is disabled as
that does not appear to work, gdb will still be built from its own,
older, split repository.