gcc: Allow gcc to be obtained from git.
This patch adds an options to the GCC version list for "GCC from git".
If this new version is selected then two new options open up to allow
the user to select a git repository URL and a repository version.
The ARC target, which was using git anyway, is now converted to use
these variables.