Enable FTPD and TELNETD in config file.
This is a slightly strange patch, and probably reflects my lack of
understanding of the kconfig system.

The options FTPD, FEATURE_FTP_WRITE, and TELNETD all have default values
of "y" set in networking/Config.in, and given these options are all not
set in the .config file I thought that the default values should apply,
and these features should all be turned on.

However, this is not the case, the features actually end up being turned
off.

I've chosen to fix this by hard coding these features to be on in the
busybox config file.  This will get things working for me, but is
probably not the best solution.